+++ src/chart/generator/plotbuilder.h
@@ -31,19 +31,24 @@ private:
 	{
 		std::size_t index{};
 		double size{};
-		const std::string *label{};
+		std::optional<std::string> label{};
 	};
 
 	struct BucketSeparationInfo
 	{
 		std::size_t itemId{};
-		std::optional<Data::SliceIndex> index{};
+		std::vector<Data::SliceIndex> index{};
 		Math::Range<> containsValues{0.0, 0.0};
 		Math::Range<> atRange{0.0, 0.0};
 
 		bool operator<(const std::size_t &id) const noexcept
 		{
 			return itemId < id;
+		}
+
+		bool operator<(const Data::SliceIndex &front) const noexcept
+		{
+			return index.front() < front;
 		}
 	};
 
@@ -53,12 +58,16 @@ private:
 	[[nodiscard]] bool linkMarkers(const Buckets &buckets,
 	    AxisId axisIndex) const;
 	void calcAxises(const Data::DataTable &dataTable,
-	    Buckets &buckets);
+	    Buckets &buckets,
+	    bool needAxisDefaultSplitMain,
+	    bool needAxisDefaultSplitSub);
 	void calcLegendAndLabel(const Data::DataTable &dataTable);
 	void calcAxis(const Data::DataTable &dataTable, AxisId type);
 	void addAlignment(const Buckets &buckets, AxisId axisIndex) const;
-	[[nodiscard]] std::vector<BucketSeparationInfo>
-	addSeparation(const Buckets &buckets, AxisId axisIndex);
+	[[nodiscard]] std::vector<BucketSeparationInfo> addSeparation(
+	    const Buckets &buckets,
+	    AxisId axisIndex,
+	    bool needAxisDefaultSplit);
 	void normalizeSizes();
 	void normalizeColors();
 	[[nodiscard]] std::vector<BucketSortInfo>
@@ -66,6 +75,8 @@ private:
 	void addSpecLayout(Buckets &buckets);
 	void addAxisLayout(Buckets &buckets,
 	    const Data::DataTable &dataTable);
+
+	[[nodiscard]] bool needAxisDefaultSplit(AxisId axisIndex) const;
 };
 }