+++ src/chart/rendering/drawaxes.cpp
@@ -2,6 +2,7 @@
 
 #include <algorithm>
 #include <cmath>
+#include <cstdint>
 #include <numbers>
 #include <optional>
 #include <ranges>
@@ -28,7 +29,7 @@
 #include "chart/main/events.h"
 #include "chart/main/style.h"
 #include "chart/options/channel.h"
-#include "dataframe/old/types.h"
+#include "chart/options/coordsystem.h"
 
 #include "drawguides.h"
 #include "drawinterlacing.h"
@@ -138,7 +139,7 @@ const DrawAxes &&DrawAxes::init() &&
 		const auto &axis = plot->axises.at(axisIndex);
 
 		const static Gen::SplitAxis::Parts oneSized{
-		    {std::nullopt, Gen::SplitAxis::Part{}}};
+		    {{}, Gen::SplitAxis::Part{}}};
 		splits[axisIndex] =
 		    axis.parts.empty() ? oneSized : axis.parts;
 
@@ -151,6 +152,9 @@ const DrawAxes &&DrawAxes::init() &&
 			auto weight = item.weight(axis.dimension.factor);
 			if (Math::Floating::is_zero(weight)) continue;
 
+			auto mainFactor = item.layer.factor(std::uint32_t{});
+			auto layer = item.layer.combine<double>();
+
 			auto needInterlacing =
 			    measEnabled == 0.0
 			    || Math::FuzzyBool::And(Math::FuzzyBool::more(weight),
@@ -159,7 +163,7 @@ const DrawAxes &&DrawAxes::init() &&
 
 			intervals.emplace_back(item.range.positive(),
 			    weight,
-			    Math::FuzzyBool::And<double>(
+			    Math::FuzzyBool::And<double>(mainFactor,
 			        Math::Niebloid::interpolate(
 			            (item.startPos ? *item.startPos
 			                           : *item.endPos)
@@ -170,6 +174,7 @@ const DrawAxes &&DrawAxes::init() &&
 			        needInterlacing),
 			    Interval::DimLabel{index,
 			        item.label,
+			        layer,
 			        !item.startPos.isAuto(),
 			        !item.endPos.isAuto()});
 
@@ -502,9 +507,7 @@ void DrawAxes::drawTitle(Gen::AxisId axisIndex,
 	}
 }
 
-void DrawAxes::drawDimensionLabels(Gen::AxisId axisIndex,
-    const Geom::AffineTransform &tr,
-    double w) const
+void DrawAxes::drawDimensionLabels(Gen::AxisId axisIndex) const
 {
 	const auto &labelStyle = rootStyle.plot.getAxis(axisIndex).label;
 
@@ -523,9 +526,7 @@ void DrawAxes::drawDimensionLabels(Gen::AxisId axisIndex,
 			drawDimensionLabel(axisIndex,
 			    origo,
 			    interval,
-			    tr,
-			    Math::FuzzyBool::And<double>(w,
-			        interval.weight,
+			    Math::FuzzyBool::And<double>(interval.weight,
 			        enabled.labels));
 		}
 	}
@@ -534,7 +535,6 @@ void DrawAxes::drawDimensionLabels(Gen::AxisId axisIndex,
 void DrawAxes::drawDimensionLabel(Gen::AxisId axisIndex,
     const Geom::Point &origo,
     const Interval &interval,
-    const Geom::AffineTransform &tr,
     double weight) const
 {
 	if (weight == 0) return;
@@ -545,7 +545,6 @@ void DrawAxes::drawDimensionLabel(Gen::AxisId axisIndex,
 	auto drawLabel = OrientedLabel{{ctx()}};
 	labelStyle.position->visit(
 	    [this,
-	        &tr,
 	        &axisIndex,
 	        &drawLabel,
 	        &labelStyle,
@@ -563,16 +562,38 @@ void DrawAxes::drawDimensionLabel(Gen::AxisId axisIndex,
 			    return;
 
 		    Geom::Point refPos;
-
+		    double spacing{};
 		    switch (position.value) {
 			    using Pos = Styles::AxisLabel::Position;
-		    case Pos::max_edge: refPos = normal; break;
+		    case Pos::max_edge:
+			    spacing =
+			        dimInfo.layer * *labelStyle.multiLevelSpacing;
+
+			    if (axisIndex == Gen::AxisId::y) {
+				    spacing *= plot->getOptions()->coordSystem.factor(
+				                   Gen::CoordSystem::polar)
+				                 * -2
+				             + 1;
+			    }
+
+			    refPos = normal;
+			    break;
 		    case Pos::axis: refPos = origo.comp(!orientation); break;
 		    default:
-		    case Pos::min_edge: refPos = Geom::Point(); break;
+		    case Pos::min_edge:
+			    spacing =
+			        -dimInfo.layer * *labelStyle.multiLevelSpacing;
+			    break;
 		    }
-
 		    auto relCenter = refPos + ident * interval.range.middle();
+		    if (axisIndex == Gen::AxisId::x
+		        && !Math::Floating::is_zero(spacing))
+			    relCenter += normal
+			               * std::copysign(
+			                   (coordSys.getOriginal(normal * spacing)
+			                       - coordSys.getOriginal({}))
+			                       .abs(),
+			                   spacing);
 
 		    auto under =
 		        labelStyle.position->interpolates()
@@ -581,18 +602,23 @@ void DrawAxes::drawDimensionLabel(Gen::AxisId axisIndex,
 		            : labelStyle.side->factor(
 		                  Styles::AxisLabel::Side::negative);
 
-		    auto draw = [&,
-		                    posDir = coordSys
-		                                 .convertDirectionAt(
-		                                     tr(Geom::Line{relCenter,
-		                                         relCenter + normal}))
-		                                 .extend(1 - 2 * under)](
-		                    const ::Anim::Weighted<bool> &str,
+		    auto posDir =
+		        coordSys
+		            .convertDirectionAt(
+		                Geom::Line{relCenter, relCenter + normal})
+		            .extend(1 - 2 * under);
+		    if (axisIndex == Gen::AxisId::y
+		        && !Math::Floating::is_zero(spacing))
+			    posDir.shift(normal * spacing);
+
+		    auto draw = [&](const ::Anim::Weighted<bool> &str,
 		                    double plusWeight = 1.0)
 		    {
 			    if (!str.value) return;
+			    auto dimValues =
+			        Gen::DimensionAxis::mergedLabels(dimInfo.index);
 			    drawLabel.draw(canvas,
-			        dimInfo.index.value,
+			        dimValues,
 			        posDir,
 			        labelStyle,
 			        0,
@@ -601,9 +627,8 @@ void DrawAxes::drawDimensionLabel(Gen::AxisId axisIndex,
 			                str.weight,
 			                plusWeight)),
 			        *rootEvents.draw.plot.axis.label,
-			        Events::Targets::dimAxisLabel(
-			            dimInfo.index.column,
-			            dimInfo.index.value,
+			        Events::Targets::dimAxisLabel(dimInfo.index,
+			            dimValues,
 			            axisIndex));
 		    };